// packages/types/src/oids.ts
export enum OIDs {
  BOOL = 16,
  BYTEA = 17,
  CHAR = 18,
  INT8 = 20,
  INT2 = 21,
  INT4 = 23,
  REGPROC = 24,
  TEXT = 25,
  OID = 26,
  TID = 27,
  XID = 28,
  CID = 29,
  JSON = 114,
  XML = 142,
  JSONARRAY = 199,
  PG_NODE_TREE = 194,
  SMGR = 210,
  POINT = 600,
  PATH = 602,
  POLYGON = 604,
  CIDR = 650,
  CIDRARRAY = 651,
  FLOAT4 = 700,
  FLOAT8 = 701,
  ABSTIME = 702,
  RELTIME = 703,
  TINTERVAL = 704,
  CIRCLE = 718,
  MACADDR8 = 774,
  MONEY = 790,
  MONEYARRAY = 791,
  MACADDR = 829,
  INET = 869,
  BOOLARRAY = 1000,
  BYTEAARRAY = 1001,
  INT2ARRAY = 1005,
  INT4ARRAY = 1007,
  REGPROCARRAY = 1008,
  TEXTARRAY = 1009,
  BPCHARARRAY = 1014,
  VARCHARARRAY = 1015,
  INT8ARRAY = 1016,
  POINTARRAY = 1017,
  FLOAT4ARRAY = 1021,
  FLOAT8ARRAY = 1022,
  OIDARRAY = 1028,
  ACLITEM = 1033,
  MACADDRARRAY = 1040,
  INETARRAY = 1041,
  BPCHAR = 1042,
  VARCHAR = 1043,
  DATE = 1082,
  TIME = 1083,
  TIMESTAMP = 1114,
  TIMESTAMPARRAY = 1115,
  DATEARRAY = 1182,
  TIMEARRAY = 1183,
  TIMESTAMPTZ = 1184,
  TIMESTAMPTZARRAY = 1185,
  INTERVAL = 1186,
  INTERVALARRAY = 1187,
  NUMERICARRAY = 1231,
  TIMETZ = 1266,
  TIMETZARRAY = 1270,
  BIT = 1560,
  VARBIT = 1562,
  NUMERIC = 1700,
  REFCURSOR = 1790,
  REGPROCEDURE = 2202,
  REGOPER = 2203,
  REGOPERATOR = 2204,
  REGCLASS = 2205,
  REGTYPE = 2206,
  UUID = 2950,
  UUIDARRAY = 2951,
  TXID_SNAPSHOT = 2970,
  PG_LSN = 3220,
  PG_NDISTINCT = 3361,
  PG_DEPENDENCIES = 3402,
  TSVECTOR = 3614,
  TSQUERY = 3615,
  GTSVECTOR = 3642,
  REGCONFIG = 3734,
  REGDICTIONARY = 3769,
  JSONB = 3802,
  JSONBARRAY = 3807,
  INT4RANGE = 3904,
  NUMRANGE = 3906,
  NUMRANGEARRAY = 3907,
  TSRANGE = 3908,
  TSTZRANGE = 3910,
  DATERANGE = 3912,
  INT8RANGE = 3926,

  REGROLE = 4096,
}